# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  httpapi.CreateServerRequest:
    properties:
      cpuMillis:
        example: 2000
        type: integer
      diskMiB:
        example: 10240
        type: integer
      displayName:
        example: Survival
        type: string
      egg:
        example: paper
        type: string
      environment:
        additionalProperties:
          type: string
        type: object
      externalTrafficPolicy:
        allOf:
        - $ref: '#/definitions/v1alpha1.TrafficPolicy'
        description: 'ExternalTrafficPolicy of the service (default Local: the server
          sees the players'' IPs).'
        enum:
        - Local
        - Cluster
      image:
        example: ghcr.io/pelican-eggs/yolks:java_21
        type: string
      loadBalancerIP:
        description: LoadBalancerIP optionally requests a fixed IP from the pool.
        example: 192.168.1.70
        type: string
      loadBalancerPool:
        description: 'LoadBalancerPool the address comes from (default: the default
          of the panel settings).'
        example: general-pool
        type: string
      memoryMiB:
        example: 4096
        type: integer
      owner:
        description: 'Owner is the username the server belongs to (default: the calling
          admin).'
        example: alice
        type: string
      ports:
        example:
        - 25565
        items:
          type: integer
        type: array
      skipInstall:
        type: boolean
      startOnCompletion:
        description: StartOnCompletion starts the server once the installation is
          finished.
        example: true
        type: boolean
      startup:
        type: string
      storageClass:
        description: 'StorageClass of the data volume (default: the default of the
          panel settings).'
        example: longhorn
        type: string
    required:
    - diskMiB
    - displayName
    - egg
    - memoryMiB
    - ports
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error:
        example: server not found
        type: string
      fields:
        additionalProperties:
          type: string
        description: Fields holds validation errors per field or variable.
        type: object
    required:
    - error
    type: object
...
  httpapi.GameServerList:
    properties:
      items:
        items:
          $ref: '#/definitions/v1alpha1.GameServer'
        type: array
    required:
    - items
    type: object
...
  httpapi.UpdateServerRequest:
    properties:
      cpuMillis:
        type: integer
      crashRestart:
        type: boolean
      diskMiB:
        type: integer
      displayName:
        type: string
      environment:
        additionalProperties:
          type: string
        type: object
      externalTrafficPolicy:
        allOf:
        - $ref: '#/definitions/v1alpha1.TrafficPolicy'
        description: ExternalTrafficPolicy of the service; admins only.
        enum:
        - Local
        - Cluster
      image:
        type: string
      loadBalancerIP:
        type: string
      loadBalancerPool:
        description: |-
          LoadBalancerPool moves the server to another enabled pool (its address changes).
          Users may change it; a fixed IP is cleared unless a new one is given.
        type: string
      memoryMiB:
        type: integer
      ports:
        items:
          type: integer
        type: array
      startup:
        type: string
      stopTimeoutSeconds:
        type: integer
    type: object
...
  v1.ManagedFieldsEntry:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the version of this resource that this field set
          applies to. The format is "group/version" just like the top-level
          APIVersion field. It is necessary to track the version of a field
          set because it cannot be automatically converted.
        type: string
      fieldsType:
        description: |-
          FieldsType is the discriminator for the different fields format and version.
          There is currently only one possible value: "FieldsV1"
        type: string
      fieldsV1:
        allOf:
        - $ref: '#/definitions/v1.FieldsV1'
        description: |-
          FieldsV1 holds the first JSON version format as described in the "FieldsV1" type.
          +optional
      manager:
        description: Manager is an identifier of the workflow managing these fields.
        type: string
      operation:
        allOf:
        - $ref: '#/definitions/v1.ManagedFieldsOperationType'
        description: |-
          Operation is the type of operation which lead to this ManagedFieldsEntry being created.
          The only valid values for this field are 'Apply' and 'Update'.
          +k8s:alpha(since: "1.37")=+k8s:required
      subresource:
        description: |-
          Subresource is the name of the subresource used to update that object, or
          empty string if the object was updated through the main resource. The
          value of this field is used to distinguish between managers, even if they
          share the same name. For example, a status update will be distinct from a
          regular update using the same manager name.
          Note that the APIVersion field is not related to the Subresource field and
          it always corresponds to the version of the main resource.
        type: string
      time:
        description: |-
          Time is the timestamp of when the ManagedFields entry was added. The
          timestamp will also be updated if a field is added, the manager
          changes any of the owned fields value or removes a field. The
          timestamp does not update when a field is removed from the entry
          because another manager took it over.
          +optional
        type: string
    type: object
...
  v1.ObjectMeta:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is an unstructured key value map stored with a resource that may be
          set by external tools to store and retrieve arbitrary metadata. They are not
          queryable and should be preserved when modifying objects.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/annotations
          +optional
        type: object
      creationTimestamp:
        description: |-
          CreationTimestamp is a timestamp representing the server time when this object was
          created. It is not guaranteed to be set in happens-before order across separate operations.
          Clients may not set this value. It is represented in RFC3339 form and is in UTC.

          Populated by the system.
          Read-only.
          Null for lists.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
          +k8s:alpha(since: "1.37")=+k8s:immutable
        type: string
      deletionGracePeriodSeconds:
        description: |-
          Number of seconds allowed for this object to gracefully terminate before
          it will be removed from the system. Only set when deletionTimestamp is also set.
          May only be shortened.
          Read-only.
          +optional
          +k8s:alpha(since: "1.37")=+k8s:optional
          +k8s:alpha(since: "1.37")=+k8s:immutable
        type: integer
      deletionTimestamp:
        description: |-
          DeletionTimestamp is RFC 3339 date and time at which this resource will be deleted. This
          field is set by the server when a graceful deletion is requested by the user, and is not
          directly settable by a client. The resource is expected to be deleted (no longer visible
          from resource lists, and not reachable by name) after the time in this field, once the
          finalizers list is empty. As long as the finalizers list contains items, deletion is blocked.
          Once the deletionTimestamp is set, this value may not be unset or be set further into the
          future, although it may be shortened or the resource may be deleted prior to this time.
          For example, a user may request that a pod is deleted in 30 seconds. The Kubelet will react
          by sending a graceful termination signal to the containers in the pod. After that 30 seconds,
          the Kubelet will send a hard termination signal (SIGKILL) to the container and after cleanup,
          remove the pod from the API. In the presence of network partitions, this object may still
          exist after this timestamp, until an administrator or automated process can determine the
          resource is fully terminated.
          If not set, graceful deletion of the object has not been requested.

          Populated by the system when a graceful deletion is requested.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
          +k8s:alpha(since: "1.37")=+k8s:optional
          +k8s:alpha(since: "1.37")=+k8s:immutable
        type: string
      finalizers:
        description: |-
          Must be empty before the object is deleted from the registry. Each entry
          is an identifier for the responsible component that will remove the entry
          from the list. If the deletionTimestamp of the object is non-nil, entries
          in this list can only be removed.
          Finalizers may be processed and removed in any order.  Order is NOT enforced
          because it introduces significant risk of stuck finalizers.
          finalizers is a shared field, any actor with permission can reorder it.
          If the finalizer list is processed in order, then this can lead to a situation
          in which the component responsible for the first finalizer in the list is
          waiting for a signal (field value, external system, or other) produced by a
          component responsible for a finalizer later in the list, resulting in a deadlock.
          Without enforced ordering finalizers are free to order amongst themselves and
          are not vulnerable to ordering changes in the list.
          +optional
          +patchStrategy=merge
          +listType=set
        items:
          type: string
        type: array
      generateName:
        description: |-
          GenerateName is an optional prefix, used by the server, to generate a unique
          name ONLY IF the Name field has not been provided.
          If this field is used, the name returned to the client will be different
          than the name passed. This value will also be combined with a unique suffix.
          The provided value has the same validation rules as the Name field,
          and may be truncated by the length of the suffix required to make the value
          unique on the server.

          If this field is specified and the generated name exists, the server will return a 409.

          Applied only if Name is not specified.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#idempotency
          +optional
        type: string
      generation:
        description: |-
          A sequence number representing a specific generation of the desired state.
          Populated by the system. Read-only.
          +optional
          +k8s:alpha(since: "1.37")=+k8s:optional
          +k8s:alpha(since: "1.37")=+k8s:minimum=0
        type: integer
      labels:
        additionalProperties:
          type: string
        description: |-
          Map of string keys and values that can be used to organize and categorize
          (scope and select) objects. May match selectors of replication controllers
          and services.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/labels
          +optional
        type: object
      managedFields:
        description: |-
          ManagedFields maps workflow-id and version to the set of fields
          that are managed by that workflow. This is mostly for internal
          housekeeping, and users typically shouldn't need to set or
          understand this field. A workflow can be the user's name, a
          controller's name, or the name of a specific apply path like
          "ci-cd". The set of fields is always in the version that the
          workflow used when modifying the object.

          +optional
          +listType=atomic
          +k8s:alpha(since: "1.37")=+k8s:optional
        items:
          $ref: '#/definitions/v1.ManagedFieldsEntry'
        type: array
      name:
        description: |-
          Name must be unique within a namespace. Is required when creating resources, although
          some resources may allow a client to request the generation of an appropriate name
          automatically. Name is primarily intended for creation idempotence and configuration
          definition.
          Cannot be updated.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace defines the space within which each name must be unique. An empty namespace is
          equivalent to the "default" namespace, but "default" is the canonical representation.
          Not all objects are required to be scoped to a namespace - the value of this field for
          those objects will be empty.

          Must be a DNS_LABEL.
          Cannot be updated.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces
          +optional
        type: string
      ownerReferences:
        description: |-
          List of objects depended by this object. If ALL objects in the list have
          been deleted, this object will be garbage collected. If this object is managed by a controller,
          then an entry in this list will point to this controller, with the controller field set to true.
          There cannot be more than one managing controller.
          +optional
          +patchMergeKey=uid
          +patchStrategy=merge
          +listType=map
          +listMapKey=uid
          +k8s:alpha(since:"1.37")=+k8s:optional
        items:
          $ref: '#/definitions/v1.OwnerReference'
        type: array
      resourceVersion:
        description: |-
          An opaque value that represents the internal version of this object that can
          be used by clients to determine when objects have changed. May be used for optimistic
          concurrency, change detection, and the watch operation on a resource or set of resources.
          Clients must treat these values as opaque and passed unmodified back to the server.
          They may only be valid for a particular resource or set of resources.

          Populated by the system.
          Read-only.
          Value must be treated as opaque by clients and .
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      selfLink:
        description: |-
          Deprecated: selfLink is a legacy read-only field that is no longer populated by the system.
          +optional
        type: string
      uid:
        description: |-
          UID is the unique in time and space value for this object. It is typically generated by
          the server on successful creation of a resource and is not allowed to change on PUT
          operations.

          Populated by the system.
          Read-only.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#uids
          +optional
          +k8s:alpha(since: "1.37")=+k8s:optional
          +k8s:alpha(since: "1.37")=+k8s:immutable
        type: string
    type: object
...
  v1alpha1.GameServer:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.GameServerSpec'
      status:
        $ref: '#/definitions/v1alpha1.GameServerStatus'
    required:
    - spec
    type: object
  v1alpha1.GameServerSpec:
    properties:
      crashRestart:
        description: |-
          CrashRestart restarts the server after a crash (not twice within 60 seconds).
          +kubebuilder:default=true
        type: boolean
      displayName:
        description: DisplayName is a human friendly name.
        type: string
      eggRef:
        description: EggRef is the name of the Egg in the same namespace.
        type: string
      environment:
        additionalProperties:
          type: string
        description: |-
          Environment holds values for the egg variables.
          +optional
        type: object
      externalTrafficPolicy:
        allOf:
        - $ref: '#/definitions/v1alpha1.TrafficPolicy'
        description: |-
          ExternalTrafficPolicy of the service; empty means Local.
          +optional
      image:
        description: Image is the runtime image (one of the egg's docker images or
          a custom one).
        type: string
      installRevision:
        description: |-
          InstallRevision triggers a (re)install whenever it differs from status.installedRevision.
          +kubebuilder:default=1
        type: integer
      loadBalancerIP:
        description: |-
          LoadBalancerIP requests a specific IP from the load balancer pool.
          +optional
        type: string
      loadBalancerPool:
        description: |-
          LoadBalancerPool is the Cilium LB IPAM pool the address comes from. The service
          gets the labels of the pool's service selector.
          +optional
        type: string
      ports:
        description: |-
          Ports are published as TCP and UDP. The first one is the primary port (SERVER_PORT).
          +kubebuilder:validation:MinItems=1
        items:
          type: integer
        type: array
      resources:
        $ref: '#/definitions/v1alpha1.Resources'
      schedules:
        description: |-
          Schedules run tasks at fixed times.
          +optional
          +listType=map
          +listMapKey=name
          +kubebuilder:validation:MaxItems=20
        items:
          $ref: '#/definitions/v1alpha1.Schedule'
        type: array
      skipInstall:
        description: |-
          SkipInstall skips the egg install script.
          +optional
        type: boolean
      startup:
        description: |-
          Startup overrides the egg's startup command.
          +optional
        type: string
      state:
        allOf:
        - $ref: '#/definitions/v1alpha1.PowerState'
        description: |-
          State is the desired power state.
          +kubebuilder:default=Stopped
      stopTimeoutSeconds:
        description: |-
          StopTimeoutSeconds is how long to wait after the stop command before killing the server.
          +kubebuilder:default=600
        type: integer
      storageClass:
        description: |-
          StorageClass of the data volume; it cannot be changed once the volume exists.
          +optional
          +kubebuilder:validation:XValidation:rule="self == oldSelf",message="storageClass cannot be changed"
        type: string
      suspended:
        description: |-
          Suspended servers are stopped and cannot be started or used by their owner (admins only).
          +optional
        type: boolean
    required:
    - eggRef
    - image
    - ports
    - resources
    type: object
  v1alpha1.GameServerStatus:
    properties:
      address:
        description: Address is the external IP assigned by the load balancer.
        type: string
      diskMeasuredAt:
        type: string
      diskUsedBytes:
        description: |-
          DiskUsedBytes is the last measured size of the server files; the files pod that
          measures it only runs on demand.
        type: integer
      installExitCode:
        description: InstallExitCode is the exit code of the last install script run.
        type: integer
      installedRevision:
        description: InstalledRevision is the last completed install revision.
        type: integer
      lastCrashAt:
        description: LastCrashAt is the time of the last detected crash.
        type: string
      lastExitCode:
        description: LastExitCode is the exit code of the last terminated game process.
        type: integer
      message:
        type: string
      observedGeneration:
        type: integer
      phase:
        $ref: '#/definitions/v1alpha1.Phase'
      podUID:
        description: PodUID identifies the current game pod.
        type: string
      restartRequired:
        description: |-
          RestartRequired is set while the running game pod was created with other runtime
          settings (image, startup, variables, memory, CPU, ports) than the spec has now.
        type: boolean
      schedules:
        description: |-
          Schedules holds the last run of every schedule.
          +optional
          +listType=map
          +listMapKey=name
        items:
          $ref: '#/definitions/v1alpha1.ScheduleStatus'
        type: array
      startedAt:
        description: StartedAt is when the current game pod was created.
        type: string
      stopRequestedAt:
        description: StopRequestedAt is set when a graceful stop was sent to the current
          pod.
        type: string
      stopTasksStartedAt:
        description: |-
          StopTasksStartedAt is set while the "stopping" schedules of the current pod run; the
          stop is sent when they are done.
          +optional
        type: string
    type: object
...
  v1alpha1.TrafficPolicy:
    enum:
    - Local
    - Cluster
    type: string
    x-enum-varnames:
    - TrafficLocal
    - TrafficCluster
...
info:
  contact: {}
  description: Game server panel for Kubernetes that runs Pterodactyl and Pelican
    eggs.
  title: Kubedactyl API
  version: dev
paths:
...
  /servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.GameServerList'
      security:
      - BearerAuth: []
      summary: List game servers
      tags:
      - Servers
    post:
      consumes:
      - application/json
      description: Creates the server and runs the egg install script. Variables are
        validated against the egg rules.
      parameters:
      - description: Server
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.CreateServerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1alpha1.GameServer'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a game server
      tags:
      - Servers
  /servers/{server}:
    delete:
      description: Deletes the server including its data volume.
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a game server
      tags:
      - Servers
    get:
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha1.GameServer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a game server
      tags:
      - Servers
    patch:
      consumes:
      - application/json
      description: Runtime settings apply on the next start. Disk can only grow.
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      - description: Changes
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1alpha1.GameServer'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a game server
      tags:
      - Servers
...
  /servers/{server}/ws:
    get:
      description: |-
        Streams events {"event": "...", "args": [...]}: "console output", "install output", "daemon
        message", "status". Send {"event":"send command","args":["say hi"]} or {"event":"set
        state","args":["start"]}.
      parameters:
      - description: Server name
        in: path
        name: server
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
      security:
      - BearerAuth: []
      summary: Open the console
      tags:
      - Servers
...
securityDefinitions:
  BearerAuth:
    description: |-
      "Bearer <token>": a session token from /auth/login or an API token (kdt_…). The web UI
      uses the kd_session cookie.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"